enable precommit
