+++ selfdrive/car/__init__.py
@@ -1,10 +1,13 @@
 # functions common among cars
 from collections import namedtuple
-from typing import Dict, List, Optional
+from enum import StrEnum
+import os
+from typing import Any, Dict, List, Optional
 
 import capnp
 
 from cereal import car
+from openpilot.common.basedir import BASEDIR
 from openpilot.common.numpy_fast import clip, interp
 
 
@@ -236,3 +239,36 @@ class CanSignalRateCalculator:
     self.previous_value = current_value
 
     return self.rate
+
+
+# interface-specific helpers
+
+INTERFACE_ATTR_FILE = {
+  "FINGERPRINTS": "fingerprints",
+  "FW_VERSIONS": "fingerprints",
+}
+
+def get_interface_attr(attr: str, combine_brands: bool = False, ignore_none: bool = False) -> Dict[str | StrEnum, Any]:
+  # read all the folders in selfdrive/car and return a dict where:
+  # - keys are all the car models or brand names
+  # - values are attr values from all car folders
+  result = {}
+  for car_folder in sorted([x[0] for x in os.walk(BASEDIR + '/selfdrive/car')]):
+    try:
+      brand_name = car_folder.split('/')[-1]
+      brand_values = __import__(f'openpilot.selfdrive.car.{brand_name}.{INTERFACE_ATTR_FILE.get(attr, "values")}', fromlist=[attr])
+      if hasattr(brand_values, attr) or not ignore_none:
+        attr_data = getattr(brand_values, attr, None)
+      else:
+        continue
+
+      if combine_brands:
+        if isinstance(attr_data, dict):
+          for f, v in attr_data.items():
+            result[f] = v
+      else:
+        result[brand_name] = attr_data
+    except (ImportError, OSError):
+      pass
+
+  return result
